Cannot delete file if used in planning

+++ public_html/php/delete_file.php
@@ -24,9 +24,20 @@ if(!isset($_POST['csrf_del'])) {
   } else {
 
     if(isset($_POST['delete'])) {
-        $target_dir = "/var/www/html/uploads/" . $_SESSION['username'] . '/' . $_POST['file_name'];
-        shell_exec('rm ' . $target_dir);
-        header("Location: ../file_upload.php?delete=Le fichier a bien été supprimé.");
+        $validation = 0;
+        $creneaux_reserves = json_decode(file_get_contents("http://10.5.0.4:5000/api/creneaux_reserves"));
+        for ($i = 0; $i < sizeof($creneaux_reserves); $i++) {
+            if ($_POST['file_name'] == $creneaux_reserves[$i][3]) {
+                $validation = 1;
+            }
+        }
+        if ($validation == 0) {
+            $target_dir = "/var/www/html/uploads/" . $_SESSION['username'] . '/' . $_POST['file_name'];
+            shell_exec('rm ' . $target_dir);
+            header("Location: ../file_upload.php?delete=Le fichier a bien été supprimé.");
+        } else {
+            header("Location: ../file_upload.php?delete=Le fichier n'a pas été supprimé car il est utilisé dans l'une de vos réservations.");
+        }
     }
 }
 ?>
